+++ src/app/page.js
@@ -1,4 +1,4 @@
-import ExchangeRateViewer from "./components/ExchangeRateViewer";
+import ExchangeRateViewer from "./components/exchange-rate-viewer";
 
 export default function Home() {
   return (
